[backend] Refactor to assertType utility
