+++ opencti-platform/opencti-graphql/src/database/engine.ts
@@ -175,6 +175,7 @@ import { DRAFT_OPERATION_CREATE, DRAFT_OPERATION_DELETE, DRAFT_OPERATION_DELETE_
 import { RELATION_SAMPLE } from '../modules/malwareAnalysis/malwareAnalysis-types';
 import { asyncMap } from '../utils/data-processing';
 import { doYield } from '../utils/eventloop-utils';
+import type { Mutable } from '../utils/type-utils';
 import { RELATION_COVERED } from '../modules/securityCoverage/securityCoverage-types';
 import type { AuthContext, AuthUser } from '../types/user';
 import type {
@@ -191,7 +192,6 @@ import type {
   StoreRelation,
 } from '../types/store';
 import type { BasicStoreSettings } from '../types/settings';
-import type { Mutable } from '../types/type-utils';
 import { completeSpecialFilterKeys } from '../utils/filtering/filtering-completeSpecialFilterKeys';
 import { IDS_ATTRIBUTES } from '../domain/attribute-utils';
 import { schemaRelationsRefDefinition } from '../schema/schema-relationsRef';
@@ -2746,11 +2746,11 @@ const buildSubQueryForFilterGroup = (
 
   const currentSubQuery = localMustFilters.length > 0
     ? {
-        bool: {
-          should: localMustFilters,
-          minimum_should_match: mode === 'or' ? 1 : localMustFilters.length,
-        },
-      }
+      bool: {
+        should: localMustFilters,
+        minimum_should_match: mode === 'or' ? 1 : localMustFilters.length,
+      },
+    }
     : null;
   return { subQuery: currentSubQuery, postFiltersTags: localPostFilterTags, resultSaltCount: localSaltCount };
 };
@@ -3123,12 +3123,12 @@ const tagFiltersForPostFiltering = (
 ) => {
   const taggedFilters: (Filter & { postFilteringTag: string })[] = filters
     ? extractFiltersFromGroup(filters, [INSTANCE_REGARDING_OF, INSTANCE_DYNAMIC_REGARDING_OF])
-        .filter((filter) => isEmptyField(filter.operator) || filter.operator === 'eq')
-        .map((filter, i) => {
-          const taggedFilter = filter as Filter & { postFilteringTag: string };
-          taggedFilter.postFilteringTag = `${i}`;
-          return taggedFilter;
-        })
+      .filter((filter) => isEmptyField(filter.operator) || filter.operator === 'eq')
+      .map((filter, i) => {
+        const taggedFilter = filter as Filter & { postFilteringTag: string };
+        taggedFilter.postFilteringTag = `${i}`;
+        return taggedFilter;
+      })
     : [];
 
   if (taggedFilters.length > 0) {
